repository page refactor + project structure change
