+++ src/dependabot/collection.py
@@ -1,0 +1,3 @@
+"""A module for managing the collection of dependabot data for the dashboard."""
+
+import streamlit as st
