+++ src/dependabot/formatting.py
@@ -1,0 +1,3 @@
+"""A module to format dependabot data for the dashboard."""
+
+import streamlit as st
