+++ src/secret_scanning/formatting.py
@@ -1,0 +1,3 @@
+"""A module to format secret scanning data for the dashboard."""
+
+import streamlit as st
